pyproject.toml
[project]
name = "epub-translate"
version = "0.1.0"
description = "Self-hosted EPUB translation web application"
requires-python = ">=3.12"
dependencies = [
  "alembic==1.14.1",
  "argon2-cffi==23.1.0",
  "beautifulsoup4==4.12.3",
  "celery[redis]==5.4.0",
  "ebooklib==0.18",
  "email-validator==2.2.0",
  "fastapi==0.115.6",
  "httpx==0.28.1",
  "jinja2==3.1.5",
  "passlib==1.7.4",
  "psycopg[binary]==3.2.13",
  "pydantic-settings==2.7.0",
  "python-multipart==0.0.20",
  "pyyaml==6.0.2",
  "redis==5.2.1",
  "sqlalchemy==2.0.36",
  "uvicorn[standard]==0.34.0",
  "itsdangerous==2.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.4",
  "pytest-asyncio==0.25.2",
  "pytest-cov==6.0.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
